split OFTPointSpriteMaterial from gltfloader
