+++ src/materials/Materials.js
@@ -14,6 +14,7 @@ export { MeshDistanceMaterial } from './MeshDistanceMaterial.js';
 export { MeshBasicMaterial } from './MeshBasicMaterial.js';
 export { MeshMatcapMaterial } from './MeshMatcapMaterial.js';
 export { OFTMatcapMaterial } from './OFTMatcapMaterial.js';
+export { OFTPointSpriteMaterial } from './OFTPointSpriteMaterial.js';
 export { LineDashedMaterial } from './LineDashedMaterial.js';
 export { LineBasicMaterial } from './LineBasicMaterial.js';
 export { Material } from './Material.js';
